+++ NEWS.md
@@ -1,3 +1,13 @@
+# influxdbr 0.14.3.9000
+
+## Bug fixes
+
+* simplifyList argument of `influx_query()` now works as expected
+
+* Encoding Fix, defaults to UTF-8 (@Avsamoht).
+
+* Empty query results yield to `NULL` instead of raising an error (#46).
+
 # influxdbr 0.14.2
 
 ## New features